Переезд navbar
